CS50P/Week_8/shirtificate_project/requirements.txt
defusedxml==0.7.1
fonttools==4.56.0
fpdf2==2.8.2
pillow==11.1.0
